Demonstrates a possible bug which causes LWS to remove the control socket and crash
The idea here was to use libwebsockets timeouts to drop clients who never
request anything (but keep the TCP session open), so they can not hold a
file descriptor for too long, making poll queue longer, etc.

To get into this situation, run an unmodified libwebsockets-test-server
(without this patch), and fire up a telnet to the address/port, and do
nothing -- the TCP session will remain open indefinitely. For
comparison, do the same to an Apache/nginx webserver -- it will drop
the connection after a while, freeing up the resources.

With this patch, the idea was to set a timer when the connection is
established at network level and removing it once a request is made (my
real application will handle timeouts after the handshake), dropping any
client who does nothing but hold a FD for too long. The problem is LWS
also times out the main control socket, closing and removing it from the
FD lookup table. In the next iteration, the library will try to index in
and crash.

I'm not sure if I'm using the timeouts it the right way, but I'm
searching a solution for this.